pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "contents-autouploader"
version = "0.1.0"
description = "Automated content creation pipeline: script generation, presentation, video, and YouTube upload"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "moonkyu"}
]
keywords = ["automation", "youtube", "presentation", "tts", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Content Creators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "anthropic>=0.18.0",
    "openai>=1.12.0",
    "python-pptx>=0.6.21",
    "moviepy>=1.0.3",
    "elevenlabs>=1.0.0",
    "google-api-python-client>=2.118.0",
    "google-auth-oauthlib>=1.2.0",
    "google-cloud-texttospeech>=2.14.0",
    "pyyaml>=6.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.2.0",
    "rich>=13.7.0",
    "pyttsx3>=2.90",
    "mutagen>=1.45.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[project.scripts]
content-uploader = "src.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
